dbt/requirements.txt
dbt==0.17.2
python-dateutil==2.8.1
PyYAML==5.3.1
sqlalchemy==1.3.20